+++ ecommerce/src/main/kotlin/it/polito/ecommerce/dto/WalletDTO.kt
@@ -3,7 +3,7 @@ package it.polito.ecommerce.dto
 import it.polito.ecommerce.domain.Wallet
 import java.math.BigDecimal
 
-data class WalletDTO(val id: Int,
+data class WalletDTO(val id: Long,
                      val balance: BigDecimal,
                      val customer: CustomerDTO)
                      //val transactionsSent: List<TransactionDTO>,
