+++ ecommerce/src/main/kotlin/it/polito/ecommerce/repositories/CustomerRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CustomerRepository : CrudRepository<Customer, Int>{
+interface CustomerRepository : CrudRepository<Customer, Long>{
 }
